tests/test_generator/test_question_service.py
import pytest
from test_generator.question_service import QuestionService  # replace with the actual import

# Mock the SQLAlchemy session
@pytest.fixture
def mock_session(mocker):
    return mocker.MagicMock()

@pytest.fixture
def retriever():
    return QuestionRetriever()

def test_get_filter_names(retriever):
    filter_names = retriever.get_filter_names()
    expected_filters = ['difficulty_levels', 'subtopic_names', 'subject_names']
    assert filter_names == expected_filters

def test__get_random_questions(retriever, mock_session):
    # Setup your mock_session to return a list of questions or whatever you expect
    questions = retriever._select_random_questions(mock_session, 10)
    assert isinstance(questions, list)
    # Add other assertions based on what you expect to be returned

def test__check_code_unique(retriever, mock_session):
    # Setup your mock_session to mimic database behavior
    mock_session.query().filter().count.return_value = 0
    is_unique = retriever._check_code_unique(mock_session, 'code123')
    assert is_unique

def test__generate_code(retriever):
    code = retriever._generate_code()
    assert len(code) == 6

# Add other test functions to test other functionalities


    def test_get_filter_names(self):
        filter_names = self.retriever.get_filter_names()
        expected_filters = ['difficulty_levels', 'subtopic_names', 'subject_names']
        self.assertListEqual(filter_names, expected_filters)

    def test__get_random_questions(self):
        # You'll probably need to set up a more elaborate mock of the session
        # to return mock Question objects with a specific structure.
        questions = self.retriever._select_random_questions(self.session, 10)
        self.assertIsInstance(questions, list)
        # other assertions based on what you expect to be returned

    def test__check_code_unique(self):
        # Setup your session mock to mimic the database behavior.
        self.session.query().filter().count.return_value = 0
        is_unique = self.retriever._check_code_unique(self.session, 'code123')
        self.assertTrue(is_unique)

    def test__generate_code(self):
        code = self.retriever._generate_code()
        self.assertEqual(len(code), 6)

    # Add other test methods to test other functionalities


if __name__ == '__main__':
    unittest.main()
